Moving my solution to a sub-module.
